Moved delete action for pool lane to its own component. Starting to refactor pool specific logic out of crown config.
